+++ example/app/app.module.ts
@@ -3,7 +3,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { LocationStrategy, HashLocationStrategy } from "@angular/common";
 import { ReactiveFormsModule, NG_VALIDATORS, FormControl } from "@angular/forms";
 import { MaterialModule } from "@angular/material";
-import { NgSemanticModule } from "ng-semantic";
 
 import { DynamicFormsCoreModule } from "@ng2-dynamic-forms/core";
 import { DynamicFormsBasicUIModule } from "@ng2-dynamic-forms/ui-basic";
@@ -12,7 +11,7 @@ import { DynamicFormsFoundationUIModule } from "@ng2-dynamic-forms/ui-foundation
 //import { DynamicFormsKendoUIModule } from "@ng2-dynamic-forms/ui-kendo";
 import { DynamicFormsMaterialUIModule } from "@ng2-dynamic-forms/ui-material";
 import { DynamicFormsPrimeNGUIModule } from "@ng2-dynamic-forms/ui-primeng";
-import { DynamicFormsSemanticUIModule } from "@ng2-dynamic-forms/ui-semantic";
+//import { DynamicFormsSemanticUIModule } from "@ng2-dynamic-forms/ui-semantic";
 
 import { BasicExampleComponent } from "./basic/basic-example.component";
 import { BootstrapExampleComponent } from "./bootstrap/bootstrap-example.component";
@@ -20,7 +19,7 @@ import { FoundationExampleComponent } from "./foundation/foundation-example.comp
 //import { KendoExampleComponent } from "./kendo/kendo-example.component";
 import { MaterialExampleComponent } from "./material/material-example.component";
 import { PrimeNGExampleComponent } from "./primeng/primeng-example.component";
-import { SemanticExampleComponent } from "./semantic/semantic-example.component";
+//import { SemanticExampleComponent } from "./semantic/semantic-example.component";
 
 import { ValidationMessageComponent } from "./validation-message/validation-message.component";
 import { AppRoutingModule } from './app.routing.module';
@@ -48,9 +47,8 @@ function testValidator(formControl: FormControl) {
         //DynamicFormsKendoUIModule,
         DynamicFormsMaterialUIModule,
         DynamicFormsPrimeNGUIModule,
-        DynamicFormsSemanticUIModule,
-        MaterialModule.forRoot(),
-        NgSemanticModule
+        //DynamicFormsSemanticUIModule,
+        MaterialModule.forRoot()
     ],
     declarations: [
         BasicExampleComponent,
@@ -59,7 +57,7 @@ function testValidator(formControl: FormControl) {
         //KendoExampleComponent,
         MaterialExampleComponent,
         PrimeNGExampleComponent,
-        SemanticExampleComponent,
+        //SemanticExampleComponent,
         ValidationMessageComponent,
         AppComponent
     ],