+++ example/systemjs.config.js
@@ -23,14 +23,14 @@
         "@angular/forms": "npm:@angular/forms/bundles/forms.umd.js",
         "@angular/forms/testing": "npm:@angular/forms/bundles/forms-testing.umd.js",
         //"@ng2-dynamic-forms": "npm:@ng2-dynamic-forms",
-         "@ng2-dynamic-forms/core": "npm:@ng2-dynamic-forms/core/bundles/core.umd.min.js",
-         "@ng2-dynamic-forms/ui-basic": "npm:@ng2-dynamic-forms/ui-basic/bundles/ui-basic.umd.min.js",
-         "@ng2-dynamic-forms/ui-bootstrap": "npm:@ng2-dynamic-forms/ui-bootstrap/bundles/ui-bootstrap.umd.min.js",
-         "@ng2-dynamic-forms/ui-foundation": "npm:@ng2-dynamic-forms/ui-foundation/bundles/ui-foundation.umd.min.js",
-         "@ng2-dynamic-forms/ui-kendo": "npm:@ng2-dynamic-forms/ui-kendo/bundles/ui-kendo.umd.min.js",
-         "@ng2-dynamic-forms/ui-material": "npm:@ng2-dynamic-forms/ui-material/bundles/ui-material.umd.min.js",
-         "@ng2-dynamic-forms/ui-primeng": "npm:@ng2-dynamic-forms/ui-primeng/bundles/ui-primeng.umd.min.js",
-         "@ng2-dynamic-forms/ui-semantic": "npm:@ng2-dynamic-forms/ui-semantic/bundles/ui-semantic.umd.min.js",
+        "@ng2-dynamic-forms/core": "npm:@ng2-dynamic-forms/core/bundles/core.umd.min.js",
+        "@ng2-dynamic-forms/ui-basic": "npm:@ng2-dynamic-forms/ui-basic/bundles/ui-basic.umd.min.js",
+        "@ng2-dynamic-forms/ui-bootstrap": "npm:@ng2-dynamic-forms/ui-bootstrap/bundles/ui-bootstrap.umd.min.js",
+        "@ng2-dynamic-forms/ui-foundation": "npm:@ng2-dynamic-forms/ui-foundation/bundles/ui-foundation.umd.min.js",
+        "@ng2-dynamic-forms/ui-kendo": "npm:@ng2-dynamic-forms/ui-kendo/bundles/ui-kendo.umd.min.js",
+        "@ng2-dynamic-forms/ui-material": "npm:@ng2-dynamic-forms/ui-material/bundles/ui-material.umd.min.js",
+        "@ng2-dynamic-forms/ui-primeng": "npm:@ng2-dynamic-forms/ui-primeng/bundles/ui-primeng.umd.min.js",
+        "@ng2-dynamic-forms/ui-semantic": "npm:@ng2-dynamic-forms/ui-semantic/bundles/ui-semantic.umd.min.js",
         "@progress": "npm:@progress",
         "@progress/kendo-angular-dropdowns": "npm:@progress/kendo-angular-dropdowns/dist/npm/main.js",
         "@progress/kendo-angular-inputs": "npm:@progress/kendo-angular-inputs/dist/npm/main.js",
@@ -70,16 +70,16 @@
             defaultExtension: "js"
         }
     };
-    /*
+
     var ng2DynamicFormsPackageNames = [
-        "@ng2-dynamic-forms/core",
-        "@ng2-dynamic-forms/ui-basic",
-        "@ng2-dynamic-forms/ui-bootstrap",
-        "@ng2-dynamic-forms/ui-foundation",
-        "@ng2-dynamic-forms/ui-kendo",
-        "@ng2-dynamic-forms/ui-material",
-        "@ng2-dynamic-forms/ui-primeng",
-        "@ng2-dynamic-forms/ui-semantic"
+        //"@ng2-dynamic-forms/core",
+        //"@ng2-dynamic-forms/ui-basic",
+        //"@ng2-dynamic-forms/ui-bootstrap",
+        //"@ng2-dynamic-forms/ui-foundation",
+        //"@ng2-dynamic-forms/ui-kendo",
+        //"@ng2-dynamic-forms/ui-material",
+        //"@ng2-dynamic-forms/ui-primeng",
+        //"@ng2-dynamic-forms/ui-semantic"
     ];
 
     ng2DynamicFormsPackageNames.forEach(function (packageName) {
@@ -89,7 +89,7 @@
             defaultExtension: "js"
         };
     });
-    */
+
     var config = {
         paths: paths,
         map: map,